-- read description --
I'm still not sure if we are supposed to be using a double solenoid, but
that is what is in the code. I spoke to mechanical and that is what they
have put on the 6CIM robot; however, on robot14 i thought we used a
single solenoid. I will leave it as is until we test / talk on
Wednesday.  I'm also using sync groups just to test them and see what
happens. In this commit I also add an information method to get some
information on dashboard and simplify some other methods.